[refactor/#47] 시험일정 ID 타입 변경

+++ core/data/application/application-api/src/main/java/com/qriz/app/core/data/application/application_api/model/Schedule.kt
@@ -4,8 +4,8 @@ import java.time.LocalDateTime
 import java.time.ZoneId
 
 data class Schedule(
-    val applicationId: Int,
-    val userApplyId: Int?,
+    val applicationId: Long,
+    val userApplyId: Long?,
     val examName: String,
     val period: String,
     val examDate: String,
